test: 인수 테스트 클래스명 변경
